Push the recursion on result/2 up to result/2 itself
This is much better - it means the reduce/2 predicate always and only defines
a single step of the small-step operational semantics, and reads far closer to
the actual inference rules in the book.

Then result/2 is only responsible for either detecting a terminating
condition, or trying to reduce another step then checking again for a
terminating condition.