Use findall/4 to do expand/3 with proper fixpoint
findall/4 [1] takes an input list (optionally uninstantiated) and returns its
results as a difference list with that list. What this means is that you get a
"free" append out of it.

In this case, that means I can make expand return the more natural value, i.e.
the union of the input set of programs with the application of all reduce
rules to that set. In turn that means the fixpoint I reach is more readable
(since I'm just checking that expansion results in the same set).

[1] http://www.swi-prolog.org/pldoc/doc_for?object=findall/4