[Ch 7] "Fully" "Working" lambda calc parser
Scare quotes because the first result this produces is indeed correct.
However, on backtracking I haven't found a way to prevent the construction of
parse trees where abstraction bodies aren't greedy. That is to say, this:

    λ.x λ.y λ.z z y x

is initially parsed, correctly, as:

    λ(x, λ(y, λ(z, ((z, y), x))))

however on backtracking it produces incorrect trees where the λ.z abstraction
doesn't capture all of the remainder, such as:

    (λ(x, (λ(y, λ(z, z)), y)), x)

The struggle continues.