[Ch 7] Force the first parse result to be the only one
This is a hack really, to address the fact that, on backtracking, my parser
starts returning non-greedy interpretations of application sequences, which in
turn means that abstraction bodies don't reach as far to the right as they
ought.

The first result is always the greedy one, however (just cos of the rule
ordering I think), so I can cut as soon as a single phrase call is successful
to ensure that this is the only result possible.

This means parse_term/2 is now deterministic, which is nice. It's still a hack
though, and I'm sure there's some way to do this properly in the DCG rules.